Move allocstats out of MemoryBank.cpp
Moved the allocstats function out of MemoryBank.cpp to remove
extra dependencies from MemoryBank.cpp and fix unit test build.
