+++ src/main/Makefile.in
@@ -31,7 +31,7 @@ SOURCES_C = complex.c inlined.c radixsort.c \
 	g_alab_her.c g_cntrlify.c g_fontdb.c g_her_glyph.c
 
 SOURCES_CXX = \
-	ArgList.cpp ArgMatcher.cpp \
+	allocstats.cpp ArgList.cpp ArgMatcher.cpp \
 	BinaryFunction.cpp Browser.cpp BuiltInFunction.cpp \
 	CellPool.cpp Closure.cpp \
 	ClosureContext.cpp CommandChronicle.cpp CommandLineArgs.cpp \
